requirements.txt
django<3
djangorestframework
psycopg2
ipython
django-extensions